Moving to an _private method
