+++ packages/workbox-core/_private.mjs
@@ -1,0 +1,5 @@
+import logger from './utils/logger.mjs';
+import WorkboxError from './models/WorkboxError.mjs';
+
+export {logger};
+export {WorkboxError};
