+++ packages/workbox-core/index.mjs
@@ -1,6 +1,5 @@
-import assert from './internal/utils/_assert.mjs';
-import logger from './internal/utils/logger.mjs';
-import WorkboxError from './internal/models/WorkboxError.mjs';
+import assert from './utils/assert.mjs';
+import * as _private from './_private.mjs';
 
 /**
  * WorkboxCore shares code across Workbox modules.
@@ -18,7 +17,6 @@ class WorkboxCore {
   }
 }
 
-export {logger};
-export {WorkboxError};
+export {_private};
 
 export default new WorkboxCore();
