+++ packages/workbox-core/models/messages/messageGenerator.mjs
@@ -1,5 +1,5 @@
-import messages from './_messages.mjs';
-import {logger} from '../../../index.mjs';
+import messages from './messages.mjs';
+import logger from '../../utils/logger.mjs';
 
 const fallback = (code, ...args) => {
   let msg = code;
