+++ packages/workbox-precaching/index.mjs
@@ -1,9 +1,10 @@
-import {WorkboxError} from 'workbox-core';
+import {_private} from 'workbox-core';
 import core from 'workbox-core';
 
 if (process.env.NODE_ENV !== 'prod') {
   if (!core.assert.isSWEnv()) {
-    throw new WorkboxError('not-in-sw', {moduleName: 'workbox-precaching'});
+    throw new _private.WorkboxError(
+      'not-in-sw', {moduleName: 'workbox-precaching'});
   }
 }
